# api/openapi.json
{
  "components": {
    "schemas": {
      "DayViewDto": {
        "properties": {
          "date": {
            "format": "date",
            "type": "string"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/EntryDto"
            },
            "type": "array"
          },
          "totals": {
            "items": {
              "$ref": "#/components/schemas/TotalDto"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
      "EntryDto": {
        "properties": {
          "amount": {
            "format": "int64",
            "type": "integer"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "loggedAt": {
            "format": "date-time",
            "type": "string"
          },
          "metricId": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "ExportDto": {
        "properties": {
          "entries": {
            "items": {
              "$ref": "#/components/schemas/ExportEntryDto"
            },
            "type": "array"
          },
          "exportedAt": {
            "format": "date-time",
            "type": "string"
          },
          "formatVersion": {
            "format": "int32",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ExportItemDto"
            },
            "type": "array"
          },
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/ExportMetricDto"
            },
            "minItems": 1,
            "type": "array"
          }
        },
        "required": [
          "entries",
          "formatVersion",
          "metrics"
        ],
        "type": "object"
      },
      "ExportEntryDto": {
        "properties": {
          "amount": {
            "format": "int64",
            "type": "integer"
          },
          "loggedAt": {
            "format": "date-time",
            "type": "string"
          },
          "metric": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "amount",
          "loggedAt",
          "metric"
        ],
        "type": "object"
      },
      "ExportItemAmountDto": {
        "properties": {
          "amount": {
            "format": "int64",
            "type": "integer"
          },
          "metric": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "amount",
          "metric"
        ],
        "type": "object"
      },
      "ExportItemDto": {
        "properties": {
          "amounts": {
            "items": {
              "$ref": "#/components/schemas/ExportItemAmountDto"
            },
            "minItems": 1,
            "type": "array"
          },
          "basisAmount": {
            "format": "int64",
            "type": "integer"
          },
          "basisUnit": {
            "minLength": 1,
            "type": "string"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "servings": {
            "items": {
              "$ref": "#/components/schemas/ExportServingDto"
            },
            "type": "array"
          }
        },
        "required": [
          "amounts",
          "basisAmount",
          "basisUnit",
          "name",
          "servings"
        ],
        "type": "object"
      },
      "ExportMetricDto": {
        "properties": {
          "canonicalUnit": {
            "minLength": 1,
            "type": "string"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "canonicalUnit",
          "name"
        ],
        "type": "object"
      },
      "ExportServingDto": {
        "properties": {
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "quantity": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "quantity"
        ],
        "type": "object"
      },
      "HelloResponse": {
        "properties": {
          "gitSha": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "ImportSummaryDto": {
        "properties": {
          "entries": {
            "format": "int32",
            "type": "integer"
          },
          "items": {
            "format": "int32",
            "type": "integer"
          },
          "metrics": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "ItemAmountDto": {
        "properties": {
          "amount": {
            "format": "int64",
            "type": "integer"
          },
          "metricId": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "amount",
          "metricId"
        ],
        "type": "object"
      },
      "ItemDto": {
        "properties": {
          "amounts": {
            "items": {
              "$ref": "#/components/schemas/ItemAmountDto"
            },
            "type": "array"
          },
          "basisAmount": {
            "format": "int64",
            "type": "integer"
          },
          "basisUnit": {
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "servings": {
            "items": {
              "$ref": "#/components/schemas/ServingDto"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
      "ItemRequest": {
        "properties": {
          "amounts": {
            "items": {
              "$ref": "#/components/schemas/ItemAmountDto"
            },
            "minItems": 1,
            "type": "array"
          },
          "basisAmount": {
            "format": "int64",
            "type": "integer"
          },
          "basisUnit": {
            "minLength": 1,
            "type": "string"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "amounts",
          "basisAmount",
          "basisUnit",
          "name"
        ],
        "type": "object"
      },
      "MetricDto": {
        "properties": {
          "canonicalUnit": {
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "name": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "NewEntryRequest": {
        "properties": {
          "amount": {
            "format": "int64",
            "type": "integer"
          },
          "loggedAt": {
            "format": "date-time",
            "type": "string"
          },
          "metricId": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "amount",
          "metricId"
        ],
        "type": "object"
      },
      "NewMetricRequest": {
        "properties": {
          "canonicalUnit": {
            "minLength": 1,
            "type": "string"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "canonicalUnit",
          "name"
        ],
        "type": "object"
      },
      "ServingDto": {
        "properties": {
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "quantity": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "ServingRequest": {
        "properties": {
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "quantity": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "quantity"
        ],
        "type": "object"
      },
      "SessionDto": {
        "properties": {
          "authenticated": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
      "TotalDto": {
        "properties": {
          "canonicalUnit": {
            "type": "string"
          },
          "metricId": {
            "format": "int64",
            "type": "integer"
          },
          "metricName": {
            "type": "string"
          },
          "total": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      }
    }
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/auth/session": {
      "get": {
        "operationId": "session",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "auth-controller"
        ]
      }
    },
    "/api/days/today": {
      "get": {
        "operationId": "today",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DayViewDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "day-controller"
        ]
      }
    },
    "/api/days/{date}": {
      "get": {
        "operationId": "day",
        "parameters": [
          {
            "in": "path",
            "name": "date",
            "required": true,
            "schema": {
              "format": "date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DayViewDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "day-controller"
        ]
      }
    },
    "/api/entries": {
      "post": {
        "operationId": "create_2",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewEntryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntryDto"
                }
              }
            },
            "description": "Created"
          }
        },
        "tags": [
          "entry-controller"
        ]
      }
    },
    "/api/entries/{id}": {
      "delete": {
        "operationId": "delete_1",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "entry-controller"
        ]
      }
    },
    "/api/export": {
      "get": {
        "operationId": "export",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ExportDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "export-import-controller"
        ]
      }
    },
    "/api/hello": {
      "get": {
        "operationId": "hello",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HelloResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "hello-controller"
        ]
      }
    },
    "/api/import": {
      "post": {
        "operationId": "importData",
        "parameters": [
          {
            "in": "query",
            "name": "mode",
            "required": true,
            "schema": {
              "enum": [
                "replace"
              ],
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ImportSummaryDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "export-import-controller"
        ]
      }
    },
    "/api/items": {
      "get": {
        "operationId": "items",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ItemDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "item-controller"
        ]
      },
      "post": {
        "operationId": "create_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ItemDto"
                }
              }
            },
            "description": "Created"
          }
        },
        "tags": [
          "item-controller"
        ]
      }
    },
    "/api/items/{id}": {
      "delete": {
        "operationId": "delete",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "item-controller"
        ]
      },
      "put": {
        "operationId": "replace",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ItemDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "item-controller"
        ]
      }
    },
    "/api/items/{id}/servings": {
      "post": {
        "operationId": "addServing",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ServingDto"
                }
              }
            },
            "description": "Created"
          }
        },
        "tags": [
          "item-controller"
        ]
      }
    },
    "/api/metrics": {
      "get": {
        "operationId": "metrics",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MetricDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "metric-controller"
        ]
      },
      "post": {
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewMetricRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MetricDto"
                }
              }
            },
            "description": "Created"
          }
        },
        "tags": [
          "metric-controller"
        ]
      }
    },
    "/api/servings/{id}": {
      "delete": {
        "operationId": "deleteServing",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "item-controller"
        ]
      },
      "put": {
        "operationId": "replaceServing",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ServingDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "item-controller"
        ]
      }
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}